Added the NoCacheLookupForPosts option.
This allows the developer to never use the previously cached value if the
incomming request is a POST. We are still caching the result for subsequent
requests (that are not POSTs). This is handy when a POST to an action is
generating a lot of computation and further GET requests would generate
the same result.